Implement steel material card
-Comment out old steel defenition
-Create new model from .inp file created with MATLAB
-Don't delete material .inp file while cleaning up
-Fix material names in sections

+++ script.py
@@ -31,12 +31,8 @@ minIncr = 1e-9
 
 #================ Materials ==================#
 # Material 1
-mat1 = "Steel"		#Material name
-mat1_Description = 'This is the description'
-mat1_dens = 8.0e-09		#Density
-mat1_E = 210000.0		#E-module
-mat1_v = 0.3			#Poisson
-mat1_yield = 355.0		#Yield stress
+matFile = 'mat_1.inp'
+mat1 = "DOMEX_S355"		#Material name
 
 # Material 2
 mat2 = "Concrete"	#Material name
@@ -144,8 +140,8 @@ simpleMonitor.printStatus(ON)
 #This makes mouse clicks into physical coordinates
 session.journalOptions.setValues(replayGeometry=COORDINATE,recoverGeometry=COORDINATE)
 
-
-mdb.Model(modelType=STANDARD_EXPLICIT, name=modelName) 	#Create a new model 
+mdb.ModelFromInputFile(name=modelName, 
+    inputFileName=matFile)
 M = mdb.models[modelName]								#For simplicity
 if len(mdb.models.keys()) > 0:							#Deletes all other models
 	a = mdb.models.items()
@@ -170,7 +166,8 @@ if 1:
 	#Delete old input files
 	inpt = glob.glob('*.inp')
 	for i in inpt:
-		os.remove(i)
+		if not i == matFile:
+			os.remove(i)
 	#Delete old jobs
 	jbs = mdb.jobs.keys()
 	if len(jbs)> 0:
@@ -185,17 +182,25 @@ if 1:
 #						MATERIALS 									 #
 #====================================================================#
 
-#================ Steel ==================#
-M.Material(description=mat1_Description, name=mat1)
-M.materials[mat1].Density(table=((mat1_dens, ), ))
-M.materials[mat1].Elastic(table=((mat1_E, mat1_v), ))
-M.materials[mat1].Plastic(table=((mat1_yield, 0.0), ))
 
-#Hardning (random linear interpolatin)
-M.materials[mat1].plastic.setValues(table=((355.0, 
-    0.0), (2000.0, 20.0)))
-#Damping (almost random mass proportional damping)
-M.materials[mat1].Damping(beta=0.0031)
+
+# #================ Steel ==================#
+# mat1_Description = 'This is the description'
+# mat1_dens = 8.0e-09		#Density
+# mat1_E = 210000.0		#E-module
+# mat1_v = 0.3			#Poisson
+# mat1_yield = 355.0		#Yield stress
+# 
+# M.Material(description=mat1_Description, name=mat1)
+# M.materials[mat1].Density(table=((mat1_dens, ), ))
+# M.materials[mat1].Elastic(table=((mat1_E, mat1_v), ))
+# M.materials[mat1].Plastic(table=((mat1_yield, 0.0), ))
+
+# #Hardning (random linear interpolatin)
+# M.materials[mat1].plastic.setValues(table=((355.0, 
+    # 0.0), (2000.0, 20.0)))
+# #Damping (almost random mass proportional damping)
+# M.materials[mat1].Damping(beta=0.0031)
 
 #================ Concrete ==================#
 M.Material(description=mat2_Description, name=mat2)
@@ -230,7 +235,7 @@ M.materials[mat3].Damping(beta=0.0031)
 #RHS 300x300
 M.BoxProfile(a=300.0, b=300.0, name='Profile-1', t1=10.0, uniformThickness=ON)
 M.BeamSection(consistentMassMatrix=False, integration=
-    DURING_ANALYSIS, material='Steel', name=sect1, poissonRatio=0.3, 
+    DURING_ANALYSIS, material=mat1, name=sect1, poissonRatio=0.3, 
     profile='Profile-1', temperatureVar=LINEAR)
 
 if imp >0:
@@ -295,7 +300,7 @@ M.IProfile(b1=300.0, b2=300.0, h=550.0, l=275.0, name=
     'Profile-2', t1=29.0, t2=29.0, t3=15.0)	#Now IPE profile, see ABAQUS for geometry definitions
 
 M.BeamSection(consistentMassMatrix=False, integration=
-    DURING_ANALYSIS, material='Steel', name=sect2, poissonRatio=0.3, 
+    DURING_ANALYSIS, material=mat1, name=sect2, poissonRatio=0.3, 
     profile='Profile-2', temperatureVar=LINEAR)
 
 #Create part
